Adding Octagon Tests

+++ Polygons.Library/IRegularPolygon.cs
@@ -9,7 +9,7 @@ namespace Polygons.Library
     internal interface IRegularPolygon
     {
          int NumberOfSides { get; set; }
-         int SideLength { get; set; }
+         double SideLength { get; set; }
 
         double GetPerimeter();
         double GetArea();
